+++ composeApp/src/commonMain/kotlin/band/effective/office/elevator/domain/phone_domain/PhoneSignIn.kt
@@ -1,4 +1,4 @@
-package band.effective.office.elevator.domain
+package band.effective.office.elevator.domain.phone_domain
 
 interface PhoneSignIn {
     fun signIn(callback: PhoneSignInResult)
