+++ composeApp/src/commonMain/kotlin/band/effective/office/elevator/domain/phone_domain/PhoneSignInResult.kt
@@ -1,4 +1,4 @@
-package band.effective.office.elevator.domain
+package band.effective.office.elevator.domain.phone_domain
 
 interface PhoneSignInResult {
     fun onSuccess(phone: String)
